sets the turbidity to 3000 when the voltage is below 2.5 V

+++ src/turbidity.h
@@ -14,7 +14,7 @@
  **/
 
 int turbidityInit(void);
-double getTurbidity(void);
+float getTurbidity(void);
 void printTurbidity(void);
 
 #endif /*TURBIDITY_H*/
